swiftvalidator/src/net/uglobal/swiftvalidator/swift/field/Field94G.java

package net.uglobal.swiftvalidator.swift.field;



/**
 * Field format : (qualifier)(address)
 * 
 * AUTO GENERATED - DO NOT HAND MODIFY
 * Generated on Sun Dec 09 17:22:28 EST 2012
 * 
 */
public class Field94G
    extends net.uglobal.swiftvalidator.swift.field.Field
{


    public Field94G(String validateData, net.uglobal.swiftvalidator.xml.beans.Field xmlField) {
        super(validateData, xmlField);
    }

    @Override
    protected void parse() {
        addComponents(net.uglobal.swiftvalidator.swift.field.FieldComponentParseUtil.parse(validateData, getContentFormat()));
    }

    public String getQualifier() {
        return getComponent1();
    }

    public String getAddress() {
        return getComponent2();
    }

    public String toString() {
        return (new StringBuilder()).append("[" ).append( "qualifier=" + getComponent1()).append( ", address=" + getComponent2() ).append( "]").toString();
    }

}
